+++ CHANGELOG.md
@@ -124,3 +124,12 @@ Released 2021-05-28
 Released 2021-06-13
 
 - Improve the messages that are logged in the entire library
+
+### Version 0.3.11
+Released 2021-06-30
+
+- Add CORS-enabled Resource class supporting:
+  - `*` 
+  - `domain.com`
+  - `*.domain.com`
+- Add Paginated Resource Helper class